requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=pypi_0
async-generator=1.10=pypi_0
attrs=20.3.0=pypi_0
backcall=0.2.0=py_0
beautifulsoup4=4.9.1=py37_0
blas=1.0=mkl
bleach=3.2.2=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.1.1=0
cached-property=1.5.2=pypi_0
certifi=2020.4.5.2=py37_0
cffi=1.14.0=py37he30daa8_1
chardet=3.0.4=py37_1003
conda=4.8.3=py37_0
conda-build=3.18.11=py37_0
conda-package-handling=1.6.1=py37h7b6447c_0
cryptography=2.9.2=py37h1ba5d50_0
cudatoolkit=10.1.243=h6bb024c_0
decorator=4.4.2=py_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
filelock=3.0.12=py_0
freetype=2.9.1=h8a8886c_1
gast=0.2.2=pypi_0
glob2=0.7=py_0
google-pasta=0.2.0=pypi_0
grpcio=1.34.0=pypi_0
h5py=3.1.0=pypi_0
icu=58.2=he6710b0_3
idna=2.9=py_1
importlib-metadata=3.3.0=pypi_0
intel-openmp=2020.1=217
ipykernel=5.4.3=pypi_0
ipython=7.15.0=py37_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.6.3=pypi_0
jedi=0.17.0=py37_0
jinja2=2.11.2=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.11=pypi_0
jupyter-console=6.2.0=pypi_0
jupyter-core=4.7.0=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libarchive=3.4.2=h62408e4_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
liblief=0.10.1=he6710b0_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_0
markdown=3.3.3=pypi_0
markupsafe=1.1.1=py37h7b6447c_0
mistune=0.8.4=pypi_0
mkl=2020.1=217
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.1.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
nbclient=0.5.1=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.2=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.3=pypi_0
ninja=1.9.0=py37hfd86e86_0
notebook=6.2.0=pypi_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
olefile=0.46=py37_0
openssl=1.1.1g=h7b6447c_0
opt-einsum=3.3.0=pypi_0
packaging=20.8=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.7.0=py_0
patchelf=0.11=he6710b0_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.1.2=py37hb39fc2d_0
pip=20.0.2=py37_3
pkginfo=1.5.0.1=py37_0
prometheus-client=0.9.0=pypi_0
prompt-toolkit=3.0.5=py_0
protobuf=3.14.0=pypi_0
psutil=5.7.0=py37h7b6447c_0
ptyprocess=0.6.0=py37_0
py-lief=0.10.1=py37h403a769_0
pycosat=0.6.3=py37h7b6447c_0
pycparser=2.20=py_0
pygments=2.6.1=py_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
pysocks=1.7.1=py37_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=pypi_0
python-libarchive-c=2.9=py_0
pytorch=1.5.1=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2020.1=py_0
pyyaml=5.3.1=py37h7b6447c_0
pyzmq=21.0.1=pypi_0
qtconsole=5.0.2=pypi_0
qtpy=1.9.0=pypi_0
readline=8.0=h7b6447c_0
requests=2.23.0=py37_0
ripgrep=11.0.2=he32d670_0
ruamel_yaml=0.15.87=py37h7b6447c_0
send2trash=1.5.0=pypi_0
setuptools=46.4.0=py37_0
six=1.14.0=py37_0
soupsieve=2.0.1=py_0
sqlite=3.31.1=h62c20be_1
tensorboard=1.15.0=pypi_0
tensorflow=1.15.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.2=pypi_0
testpath=0.4.4=pypi_0
tk=8.6.8=hbc83047_0
torchvision=0.6.1=py37_cu101
tornado=6.1=pypi_0
tqdm=4.46.0=py_0
traitlets=4.3.3=py37_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.25.8=py37_0
wcwidth=0.2.4=py_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yaml=0.1.7=had09818_2
zipp=3.4.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
